addition of start screen
