+++ app/src/main/java/com/example/touristguide/MainActivity.java
@@ -1,31 +1,24 @@
 package com.example.touristguide;
 import androidx.fragment.app.FragmentActivity;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
+import android.content.Intent;
 import android.os.Bundle;
-import com.google.android.gms.maps.GoogleMap;
-import com.google.android.gms.maps.OnMapReadyCallback;
-import com.google.android.gms.maps.SupportMapFragment;
-public class MainActivity extends FragmentActivity {
+import android.view.View;
+import android.widget.Button;
 
-    @Override
+public class MainActivity extends FragmentActivity implements View.OnClickListener {
+
+    Button startButton ;
+
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        // Initialize the SupportMapFragment
-        FragmentManager fragmentManager = getSupportFragmentManager();
-        FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
-        SupportMapFragment mapFragment = new SupportMapFragment();
-        fragmentTransaction.add(R.id.mapFragment, mapFragment);
-        fragmentTransaction.commit();
-
-        // Set up the Google Map
-        mapFragment.getMapAsync(new OnMapReadyCallback() {
-            @Override
-            public void onMapReady(GoogleMap googleMap) {
-                // You can customize the map and add markers, etc., here.
-            }
-        });
+        setContentView(R.layout.start_screen); // Set your layout XML file
+        startButton = findViewById(R.id.startButton);
+        startButton.setOnClickListener(this); // Set the click listener here
     }
+    @Override
+    public void onClick(View v) {
+        Intent intent = new Intent(v.getContext(), MapsActivity.class); // Replace MainActivity.class with the actual target class
+        startActivity(intent);
+    }
+
 }
