+++ app/src/main/java/com/example/touristguide/MapsActivity.java
@@ -1,0 +1,35 @@
+package com.example.touristguide;
+
+import android.os.Bundle;
+
+
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.OnMapReadyCallback;
+import com.google.android.gms.maps.SupportMapFragment;
+
+public class MapsActivity extends FragmentActivity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        // Initialize the SupportMapFragment
+        FragmentManager fragmentManager = getSupportFragmentManager();
+        FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+        SupportMapFragment mapFragment = new SupportMapFragment();
+        fragmentTransaction.add(R.id.mapFragment, mapFragment);
+        fragmentTransaction.commit();
+
+        // Set up the Google Map
+        mapFragment.getMapAsync(new OnMapReadyCallback() {
+            @Override
+            public void onMapReady(GoogleMap googleMap) {
+                // You can customize the map and add markers, etc., here.
+            }
+        });
+    }
+}
