+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">Tourist Guide</string>
+    <string name="start">Start</string>
 </resources>
